# feat(matrices): use scc to speed up charpoly