@oscarbenjamin 2023-07-20
This is ready for review. I've added a release note.